move pull slip bibInfo to BibSection component files
